refactor(healthcheck): adds testing improvements
