+++ src/routes/health/health.spec.ts
@@ -1,5 +1,6 @@
 import { handler as healthCheck, Health, setRouteHealth } from '.';
 import { Request, Response } from 'express';
+import { router } from '../index';
 
 describe('API Routing Health Check', () => {
   let req: Partial<Request>;
@@ -18,11 +19,21 @@ describe('API Routing Health Check', () => {
     status = Health.busy;
     req.path = '/init';
     setRouteHealth(status, <Request>req);
+    let result;
+    // tslint:disable: no-any no-unsafe-any
+    router.stack.forEach(endpoint => {
+      if (endpoint.route.path === req.path) {
+        result = endpoint.routeHealth;
+        // tslint:enable: no-any no-unsafe-any
+      }
+    });
+    expect(result).toBe(Health.busy);
   });
 
   it('returns a response', () => {
     healthCheck(<Request>req, <Response>res);
     expect(res.send).toBeCalledTimes(1);
     expect(res.status).toHaveBeenCalledWith(200);
+    expect.assertions(2);
   });
 });
