+++ src/routes/init/init.spec.ts
@@ -1,6 +1,6 @@
 import { handler as dbInit } from './index';
 import { Request, Response } from 'express';
-
+jest.mock('../health');
 jest.mock('./db/index', () => ({
   importDbEntries: jest
     .fn()
@@ -9,16 +9,6 @@ jest.mock('./db/index', () => ({
     .mockRejectedValueOnce(new Error('I am Error.'))
 }));
 
-jest.mock('../health', () => ({
-  setRouteHealth: jest.fn().mockResolvedValue('{}'),
-  Health: jest.fn().mockImplementation(() => {
-    enum Health {
-      good = 'OK',
-      bad = 'This resource is either not running or not working. Better fix it.',
-      busy = 'This resource is busy, please try again later.'
-    }
-  })
-}));
 
 describe('init', () => {
   let req: Partial<Request>;
