+++ src/routes/pdf/pdf.spec.ts
@@ -1,16 +1,6 @@
 import { handler as pdfDownload } from '.';
 import { Request, Response } from 'express';
-
-jest.mock('../health', () => ({
-  setRouteHealth: jest.fn().mockResolvedValue('{}'),
-  Health: jest.fn().mockImplementation(() => {
-    enum Health {
-      good = 'OK',
-      bad = 'This resource is either not running or not working. Better fix it.',
-      busy = 'This resource is busy, please try again later.'
-    }
-  })
-}));
+jest.mock('../health');
 
 describe('PDF generation middleware function', () => {
   it('returns a response', () => {
